mv atomEffect2.ts to atomEffect.ts
