+++ src/index.ts
@@ -1,2 +1,1 @@
 export { atomEffect } from './atomEffect'
-export { atomEffect as atomEffect2 } from './atomEffect2'
